+++ packages/sdk/src/index.ts
@@ -22,6 +22,7 @@ import { keepkeyBexWallet } from "@swapkit/wallets/keepkey-bex";
 import { keplrWallet } from "@swapkit/wallets/keplr";
 import { keystoreWallet } from "@swapkit/wallets/keystore";
 import { ledgerWallet } from "@swapkit/wallets/ledger";
+import { walletSelectorWallet } from "@swapkit/wallets/near-wallet-selector";
 import { okxWallet } from "@swapkit/wallets/okx";
 import { onekeyWallet } from "@swapkit/wallets/onekey";
 import { phantomWallet } from "@swapkit/wallets/phantom";
@@ -74,6 +75,7 @@ export {
   tronlinkWallet,
   vultisigWallet,
   walletconnectWallet,
+  walletSelectorWallet,
   xamanWallet,
 };
 
@@ -110,6 +112,7 @@ export const defaultWallets = {
   ...tronlinkWallet,
   ...vultisigWallet,
   ...walletconnectWallet,
+  ...walletSelectorWallet,
   ...xamanWallet,
 };