+++ packages/toolboxes/src/near/types/toolbox.ts
@@ -1,51 +1,10 @@
-import type {
-  AssetValue,
-  ChainSigner,
-  DerivationPathArray,
-  GenericCreateTransactionParams,
-  GenericTransferParams,
-} from "@swapkit/helpers";
-import type { Account, Contract, KeyPairSigner, providers, Signer, transactions } from "near-api-js";
-import type { Action, SignedTransaction, Transaction } from "near-api-js/lib/transaction";
+import type { Account, Contract } from "@near-js/accounts";
+import type { JsonRpcProvider } from "@near-js/providers";
+import type { Action, SignedTransaction, Transaction } from "@near-js/transactions";
+import type { AssetValue, DerivationPathArray } from "@swapkit/helpers";
 import type { NEP141Token } from "../helpers/nep141";
+import type { NearCreateTransactionParams, NearFunctionCallParams, NearSigner, NearTransferParams } from "../types";
 import type { NearContractInterface, NearGasEstimateParams } from "../types/contract";
-
-interface NearKeyPairSigner
-  extends KeyPairSigner,
-    Omit<ChainSigner<typeof transactions.Transaction, typeof transactions.SignedTransaction>, "signTransaction"> {}
-
-interface NearGeneralSigner
-  extends Signer,
-    Omit<ChainSigner<typeof transactions.Transaction, typeof transactions.SignedTransaction>, "signTransaction"> {}
-
-export type NearSigner = NearKeyPairSigner | NearGeneralSigner;
-
-export type NearToolboxParams =
-  | { signer?: NearSigner; accountId?: string }
-  | { phrase?: string; index?: number; derivationPath?: DerivationPathArray };
-
-export interface NearTransferParams extends GenericTransferParams {}
-
-export interface NearConfig {
-  networkId: "mainnet" | "testnet" | "betanet";
-  nodeUrl: string;
-  walletUrl?: string;
-  helperUrl?: string;
-  keyStore?: any;
-}
-
-export interface NearFunctionCallParams {
-  contractId: string;
-  methodName: string;
-  args: Uint8Array | Record<string, any>;
-  deposit?: bigint | string | number;
-  gas?: bigint | string | number;
-}
-
-export interface NearCreateTransactionParams extends Omit<GenericCreateTransactionParams, "feeRate"> {
-  attachedDeposit?: string;
-  functionCall?: { methodName: string; args: object; attachedDeposit: string; gas: string; contractId: string };
-}
 
 export interface BatchTransaction {
   receiverId: string;
@@ -72,15 +31,13 @@ export interface GetSignerFromPhraseParams {
 export interface NearToolbox {
   getAddress: () => Promise<string>;
   getPublicKey: () => Promise<string>;
-  provider: providers.JsonRpcProvider;
+  provider: typeof JsonRpcProvider;
   transfer: (params: NearTransferParams) => Promise<string>;
   createAction: (params: CreateActionParams) => Promise<Action>;
   createTransaction: (params: NearCreateTransactionParams) => Promise<Transaction>;
   createContractFunctionCall: (params: ContractFunctionCallParams) => Promise<Transaction>;
   estimateTransactionFee: (params: NearTransferParams | NearGasEstimateParams) => Promise<AssetValue>;
   broadcastTransaction: (signedTransaction: SignedTransaction) => Promise<string>;
-  signAndBroadcastTransaction: (transaction: Transaction) => Promise<string>;
-  signer: NearSigner | undefined;
   signTransaction: (transaction: Transaction) => Promise<SignedTransaction>;
   getBalance: (address: string) => Promise<AssetValue[]>;
   validateAddress: (address: string) => boolean;
@@ -94,4 +51,5 @@ export interface NearToolbox {
   getGasPrice: () => Promise<string>;
   estimateGas: (params: NearGasEstimateParams, account?: Account) => Promise<AssetValue>;
   serializeTransaction: (params: Transaction) => Promise<string>;
+  signAndSendTransaction: (params: Transaction) => Promise<string>;
 }