+++ packages/ui/src/react/index.tsx
@@ -1,4 +1,10 @@
-export * from "./components/asset-input";
-export * from "./components/connect-button";
-export * from "./context";
+export * from "./components/asset-icon";
+export * from "./components/chain-icon";
+export * from "./components/composable/swap-asset-item";
+export * from "./components/composable/swap-asset-select";
+export * from "./components/composable/swap-input-chain-selector";
+export * from "./components/wallet-icon";
+
+export * from "./swapkit-context";
 export * from "./swapkit-widget";
+export * from "./types";