+++ packages/browser/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @swapkit/browser
 
+## 4.0.38
+
+### Patch Changes
+
+- Updated dependencies [[`68cb4ce`](https://github.com/swapkit/SwapKit/commit/68cb4cef9ee3a4451a44eea45f52dca34506b973)]:
+  - @swapkit/helpers@4.3.2
+  - @swapkit/plugins@4.1.6
+  - @swapkit/wallets@4.2.2
+  - @swapkit/core@4.0.36
+
+## 4.0.37
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/helpers@4.3.1
+  - @swapkit/core@4.0.35
+  - @swapkit/plugins@4.1.5
+  - @swapkit/wallets@4.2.1
+
+## 4.0.36
+
+### Patch Changes
+
+- Updated dependencies [[`357a262`](https://github.com/swapkit/SwapKit/commit/357a2625fe7d528b377b2cac954dcab7a6adc421), [`0d56165`](https://github.com/swapkit/SwapKit/commit/0d56165668372c41711bb3fbea21436cb1c7bbec)]:
+  - @swapkit/helpers@4.3.0
+  - @swapkit/wallets@4.2.0
+  - @swapkit/core@4.0.34
+  - @swapkit/plugins@4.1.4
+
+## 4.0.35
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/core@4.0.33
+  - @swapkit/helpers@4.2.3
+  - @swapkit/plugins@4.1.3
+  - @swapkit/wallets@4.1.31
+
+## 4.0.34
+
+### Patch Changes
+
+- Updated dependencies [[`41ea03c`](https://github.com/swapkit/SwapKit/commit/41ea03cc82888137680b97f1fba986f5750692c2)]:
+  - @swapkit/helpers@4.2.2
+  - @swapkit/core@4.0.32
+  - @swapkit/plugins@4.1.2
+  - @swapkit/wallets@4.1.30
+
 ## 4.0.33
 
 ### Patch Changes